refactor(analysis): decompose recipe service and lazy-load barrels
