+++ src/flameox/adapters/__init__.py
@@ -1,53 +1,40 @@
-from flameox.adapters.benchmark_samples import (
-    BenchmarkSamplesExtractionResult,
-    BenchmarkSamplesExtractor,
-    BenchmarkSamplesV1,
-)
-from flameox.adapters.client_setup import (
-    ALL_SETUP_CLIENTS,
-    ClientConfigEdit,
-    ClientConfigRegistry,
-    ClientPlanAction,
-    Launcher,
-    SetupClient,
-)
-from flameox.adapters.coverage import CoverageExtractionResult, CoverageExtractor
-from flameox.adapters.memray import MemrayExtractionResult, MemrayExtractor
-from flameox.adapters.nsight_systems import (
-    NsightSystemsExtractionResult,
-    NsightSystemsExtractor,
-)
-from flameox.adapters.observations import (
-    ObservationExtractionResult,
-    ObservationExtractor,
-)
-from flameox.adapters.perfetto import (
-    PerfettoExtractionResult,
-    PerfettoExtractor,
-    TraceEvent,
-    TraceWindowResult,
-)
-from flameox.adapters.pyperf import PyPerfExtractionResult, PyPerfExtractor
-from flameox.adapters.pytest import PytestExtractionResult, PytestExtractor
-from flameox.adapters.python_startup import (
-    PythonStartupExtractionResult,
-    PythonStartupExtractor,
-)
-from flameox.adapters.registry import (
-    AdapterApproval,
-    AdapterDescriptor,
-    AdapterDiscoveryResult,
-    AdapterRegistry,
-)
-from flameox.adapters.setup_runtime import (
-    ManagedRuntime,
-    RuntimeInstallation,
-    TraceProcessorInstallation,
-    install_trace_processor,
-)
-from flameox.adapters.torch_profiler import (
-    TorchProfilerCaptureOptions,
-    TorchProfilerSchedule,
+from __future__ import annotations
+
+from importlib import import_module
+from typing import TYPE_CHECKING
+
+# The conditional imports preserve static types while runtime exports stay lazy.
+# ruff: noqa: F405
+if TYPE_CHECKING:
+    from flameox.adapters.benchmark_samples import *  # noqa: F403
+    from flameox.adapters.client_setup import *  # noqa: F403
+    from flameox.adapters.coverage import *  # noqa: F403
+    from flameox.adapters.memray import *  # noqa: F403
+    from flameox.adapters.nsight_systems import *  # noqa: F403
+    from flameox.adapters.observations import *  # noqa: F403
+    from flameox.adapters.perfetto import *  # noqa: F403
+    from flameox.adapters.pyperf import *  # noqa: F403
+    from flameox.adapters.pytest import *  # noqa: F403
+    from flameox.adapters.python_startup import *  # noqa: F403
+    from flameox.adapters.registry import *  # noqa: F403
+    from flameox.adapters.setup_runtime import *  # noqa: F403
+    from flameox.adapters.torch_profiler import *  # noqa: F403
+
+
+_MODULES = (
+    "benchmark_samples",
+    "client_setup",
+    "coverage",
+    "memray",
+    "nsight_systems",
+    "observations",
+    "perfetto",
+    "pyperf",
+    "pytest",
+    "python_startup",
+    "registry",
+    "setup_runtime",
+    "torch_profiler",
 )
 
 __all__ = [
@@ -89,3 +76,21 @@ __all__ = [
     "TraceWindowResult",
     "install_trace_processor",
 ]
+
+
+def __getattr__(name: str) -> object:
+    if name not in __all__:
+        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+    for module_name in _MODULES:
+        module = import_module(f"{__name__}.{module_name}")
+        try:
+            value = getattr(module, name)
+        except AttributeError:
+            continue
+        globals()[name] = value
+        return value
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+def __dir__() -> list[str]:
+    return sorted(set(globals()) | set(__all__))
